requirements.txt
openai
python-dotenv
python-environ
inquirer
SQLAlchemy
beautifulsoup4
markdown
markdown-checklist
autopep8
termcolor
pyyaml
pytest
pytest-env
progressbar2
cuid
tiktoken
pytz
fastapi
uvicorn[standard]
redis
fakeredis
setproctitle
# Flask
# Flask-Cors
# gpt4all
# apscheduler